moved all wrap method & edit dashboard
