+++ wrap/templates/wrap/dashboard.html
@@ -20,16 +20,9 @@
   <div class="Wrap">
     <div class="WrapTemp">
       <h2>Wrap Name</h2>
-        <a href="{% url 'your_wrap' %}">
+        <a href="{% url 'view_wraps' %}">
         <button type="button" class="viewBTN"> View </button>
         </a>
-      <button class="deleteBTN"> Delete </button>
-    </div>
-    <div class="WrapTemp">
-      <h2> Wrap Name </h2>
-        <a href="{% url 'your_wrap' %}">
-        <button type="button" class="viewBTN"> View </button>
-        </a>      <button class="deleteBTN"> Delete </button>
     </div>
   </div>
   <button class="createBTN"> Create </button>
